Updated ReadMe for configuring build path
I found that when I cloned this my build path was mapped to a location on an F:/ drive... which I do not have. I had to configure the build path for my dependencies in the libs directory. Decided it best to share this in case others needed the quick start info